Добавлены первые данные, добавлено окно для второй таблицы

+++ Models/Passengers.cs
@@ -8,5 +8,5 @@ public class Passengers
     public int Age { get; set; }
     public string Nationality { get; set; }
     public int Passport { get; set; }
-    public int Contact { get; set; }
+    public long Contact { get; set; }
 }
